internal: move linux specific code to separate package
Move some linux specific code out of package internal into
package linux. As a rule of thumb, code probably belongs in
this package if it calls into package unix.
