+++ internal/linux/doc.go
@@ -1,0 +1,2 @@
+// Package linux contains OS specific wrappers around package unix.
+package linux
