+++ internal/linux/statfs.go
@@ -1,4 +1,4 @@
-package internal
+package linux
 
 import (
 	"unsafe"
