+++ internal/linux/statfs_test.go
@@ -1,4 +1,4 @@
-package internal
+package linux
 
 import (
 	"testing"
