+++ internal/linux/testdata/auxv64le_no_vdso.bin
