+++ internal/linux/testdata/vdso.bin
